requirements.txt
-e git+https://github.com/aquavitae/aafigure.git#egg=aafigure-master
alabaster==0.7.4
Babel==1.3
coverage==3.7.1
docutils==0.12
Jinja2==2.7.3
MarkupSafe==0.23
nose==1.3.7
pdfrw==0.1
Pillow==2.8.1
Pygments==2.0.2
PyPDF2==1.24
pytz==2015.4
reportlab==3.2.0
rson==0.9
six==1.9.0
smartypants==1.8.6
snowballstemmer==1.2.0
Sphinx==1.3.1
sphinx-rtd-theme==0.1.8
-e git+https://github.com/aquavitae/svg2rlg.git#egg=svg2rlg-master
Tenjin==1.1.1
